+++ src/Application/Components/Step/StepPlan.php
@@ -1,13 +1,5 @@
 <?php
 
-/*
- * This file is part of the Proximum Vimeet project.
- *
- * Copyright (C) Proximum
- *
- * @author Elao <contact@elao.com>
- */
-
 namespace Proximum\Vimeet\Application\Components\Step;
 
 use Proximum\Vimeet\Application\Command\Package\Step\SelectPlan;
@@ -16,16 +8,9 @@ use Proximum\Vimeet\Domain\Model\Sheet;
 
 class StepPlan
 {
-    /**
-     * @var CartManager
-     */
+    /** @var CartManager */
     private $cartManager;
 
-    /**
-     * StepPlan constructor.
-     *
-     * @param CartManager $cartManager
-     */
     public function __construct(CartManager $cartManager)
     {
         $this->cartManager = $cartManager;
@@ -37,10 +22,10 @@ class StepPlan
      *
      * @return SelectPlan
      */
-    public function build(Sheet $sheet, $stepIndex)
+    public function build(Sheet $sheet, $stepIndex): SelectPlan
     {
-        $command      = new SelectPlan($sheet, $stepIndex);
-        $cart         = $this->cartManager->getCart($command->sheet, $command->currentStep);
+        $command = new SelectPlan($sheet, $stepIndex);
+        $cart = $this->cartManager->getCart($command->sheet, $command->currentStep);
         $selectedPlan = $cart->getPlanRow();
 
         if (null !== $selectedPlan) {