+++ tests/Application/Query/Sheet/PaginatedSheetListViewQueryHandlerTest.php
@@ -1,13 +1,5 @@
 <?php
 
-/*
- * This file is part of the Proximum Vimeet project.
- *
- * Copyright (C) Proximum
- *
- * @author Elao <contact@elao.com>
- */
-
 namespace Proximum\Vimeet\Tests\Application\Query\Sheet;
 
 use PHPUnit\Framework\TestCase;
@@ -37,7 +29,7 @@ use Proximum\Vimeet\Infrastructure\Bundle\InfrastructureBundle\Security\Imperson
 
 class PaginatedSheetListViewQueryHandlerTest extends TestCase
 {
-    public function testHandle()
+    public function testHandle(): void
     {
         $event = $this->prophesize(Event::class);
         $admin = $this->prophesize(Admin::class);
@@ -81,7 +73,7 @@ class PaginatedSheetListViewQueryHandlerTest extends TestCase
         $sheet1->getValidationState()->willReturn('validationState1');
         $sheet1->getCompleteness()->willReturn(45);
         $sheet1->isEnabled()->willReturn(true);
-        $sheet1->isInCatalog()->willReturn(true);
+        $sheet1->isInInternalCatalog()->willReturn(true);
         $sheet1->isAccepted()->willReturn(true);
         $sheet1->isValidated()->willReturn(false);
         $sheet1->attend()->willReturn(true);
@@ -98,7 +90,7 @@ class PaginatedSheetListViewQueryHandlerTest extends TestCase
         $type2->getCategoriesTitles('fr')->willReturn([]);
         $type2->getTitle('fr')->willReturn('type2');
         $sheet1->getType()->willReturn($type1->reveal());
-        $sheet1->countParticipant()->willReturn(1);
+        $sheet1->countParticipants()->willReturn(1);
         $sheet1->getGroup()->willReturn($group->reveal());
         $sheet1->getTraceableName()->willReturn('Sheet');
         $group->getTitle()->willReturn('group title 1');
@@ -112,9 +104,9 @@ class PaginatedSheetListViewQueryHandlerTest extends TestCase
         $sheet2->getValidationState()->willReturn('validationState2');
         $sheet2->getCompleteness()->willReturn(75);
         $sheet2->isEnabled()->willReturn(false);
-        $sheet2->isInCatalog()->willReturn(false);
+        $sheet2->isInInternalCatalog()->willReturn(false);
         $sheet2->getType()->willReturn($type2->reveal());
-        $sheet2->countParticipant()->willReturn(2);
+        $sheet2->countParticipants()->willReturn(2);
         $sheet2->getGroup()->willReturn(null);
         $sheet2->getSpot()->willReturn(null);
         $sheet2->isAccepted()->willReturn(false);
@@ -220,6 +212,6 @@ class PaginatedSheetListViewQueryHandlerTest extends TestCase
             $expectedSheet2,
         ], 1, 20, 2);
 
-        $this->assertEquals($expected, $result);
+        self::assertEquals($expected, $result);
     }
 }